+++ types/managed-role.ts
@@ -3,3 +3,8 @@ export interface RoleData {
   roleName: string;
   description: string;
 }
+
+export interface EnroleApiParams {
+  roleName: string;
+  description: string;
+}
